client/requirements.txt
aiohttp==3.7.4
azure-functions==1.6.0